Add pinctrl-fake-gpio-toggler.
This change adds a per-pinctrl_fake_gpio_chip work item that toggles the value of any input GPIO periodically.

The default toggling period is 1 Hz (i.e. gpio input is inverted once per second). Thus, the period of the input signal, when viewed as a square wave, is 0.5 Hz (it takes 2 seconds to go from high, to low, and back to high).

Additional work that needs to be done, is:
a) add a debugfs interface to dynamically adjust the toggle period, per-pin
b) trigger the GPIO interrupt mechanism that notifies sysfs of GPIO changes.