Call pinctrl_fake_irq_handler() directly from scheduled work.
It wasn't particularly obvious how to trigger a fake GPIO interrupt via software, so I called pinctrl_fake_irq_handler() directly with the appropriate arguments.

It would probably be wise to introduce a 'pended' array to indicate which pins have an interrupt triggered, so that batches of them can be handled from wihtin the irq_handler().

Another improvement would be to use arrays of longs and the Linux kernel's bitwise macros to store the majority of the gpio data. This would save a fair bit of memory from being wasted.