Fixup for pinctrl-fake debugging messages.
With a recent gpio-related change made upstream, the cdev field of struct
gpio_chip was moved into struct gpio_dev gpiodev, which is opaque.

This is a dirty hack to get the definition of struct gpio_dev.

ATM, it's only used for dev_info, dev_err, and dev_dbg, and therefore is not
100% necessary. It will be removed in a future revision.