gpio-fake module parsing and probing with devicetree
I moved the fake gpio functionality out of the pinctrl-fake driver directory
and into the gpio driver directory, which is where it belongs. When gpio-fake
is probed, I have access to the parent device platform data, which is the
parent pinctrl-fake instance.

[ 1567.058154] pinctrl-fake pinctrl: Added pinctrl_fake @ ffff88003cf1f3c0, pdev @ ffff88003e1bca00, dev @ ffff88003e1bca10
[ 1576.039909] Fake Gpio Driver: Copyright (c) 2016, Christopher Friedt
[ 1576.039946] gpio-fake pinctrl:gpio@0: in gpio_fake_probe()
[ 1576.039949] gpio-fake pinctrl:gpio@0: dev: ffff88003d6a0210, parent: ffff88003e1bca10, pctrl: ffff88003cf1f3c0

I would like to now ensure that I can parse the gpio devicetree data and
instantiate all of the required gpio pins for a given gpio-chip.