Added code to handle pinctrl-fake-mappings, discoball.dts changes.
Previously pinctrl_register_mappings() was not an exported symbol, so could
not be called from module code. Fixed that.

Unfortunately, there is no pinctrl_unregister_mappings(), which explains why
there are triplicates of all of my mappings m(. Will add that when upstreaming.

I would like pinctrl to handle my GPIOs if possible, but it will take a bit
more tinkering to get that to a useable state. I might not have needed
gpio-fake.c at all, but who knows..