gpio-fake and pinctrl-fake both working via devicetree
Previously, the gpio-fake driver lived with the pinctrl-fake driver and the
number of gpio-fake devices, the number of pins, etc, were all hard-coded and
set at compile time. Since developers can now describe the hardware present on
a given system using devicetree, in a way that is invariant wrt the OS kernel,
I was able to abstract the pins available to pinctrl-fake, the grouping, the
devices, names, etc, all out to a .dts file. DT firmware files are even
dynamically reloadable now, fwiu, so this system for faking out bus drivers is
really quite flexible.

I'm tempted to bring control for the GPIO (and possibly other IIO devices such
as DACs or ADCs) out to userspace too.